Fix Mac OS compile error.
