+++ engine/src/lib.rs
@@ -296,3 +296,363 @@ pub(crate) mod freetype_sys_patch {
     }
     pub(crate) type FT_SfntName = FT_SfntName_;
 }
+
+#[cfg(target_os = "macos")]
+pub(crate) mod cf_prelude {
+    pub(crate) use core_foundation::{
+        array::{
+            kCFTypeArrayCallBacks, CFArrayCallBacks, CFArrayCreate, CFArrayGetCount, CFArrayRef,
+            __CFArray,
+        },
+        attributed_string::{CFAttributedString, CFAttributedStringCreate, CFAttributedStringRef},
+        base::{
+            kCFAllocatorDefault, kCFAllocatorNull, CFAllocatorRef, CFComparisonResult, CFEqual,
+            CFHashCode, CFIndex, CFOptionFlags, CFRange, CFRelease, CFRetain, CFTypeRef, ToVoid,
+        },
+        boolean::{kCFBooleanTrue, CFBooleanRef},
+        dictionary::{
+            kCFTypeDictionaryKeyCallBacks, kCFTypeDictionaryValueCallBacks, CFDictionary,
+            CFDictionaryAddValue, CFDictionaryCopyDescriptionCallBack, CFDictionaryCreate,
+            CFDictionaryCreateMutable, CFDictionaryEqualCallBack, CFDictionaryGetValueIfPresent,
+            CFDictionaryHashCallBack, CFDictionaryKeyCallBacks, CFDictionaryRef,
+            CFDictionaryReleaseCallBack, CFDictionaryRetainCallBack, CFDictionaryValueCallBacks,
+            CFMutableDictionaryRef,
+        },
+        number::{CFNumberCompare, CFNumberCreate, CFNumberGetValue, CFNumberRef, CFNumberType},
+        set::{kCFTypeSetCallBacks, CFSetCreate, CFSetRef},
+        string::{
+            kCFStringEncodingUTF8, CFStringCompareFlags, CFStringCreateWithBytes,
+            CFStringCreateWithCString, CFStringEncoding, CFStringGetCString, CFStringGetLength,
+            CFStringRef,
+        },
+        url::{CFURLGetFileSystemRepresentation, CFURLRef},
+    };
+    pub(crate) const kCFNumberMaxType: CFNumberType = 16;
+    pub(crate) const kCFNumberCGFloatType: CFNumberType = 16;
+    pub(crate) const kCFNumberNSIntegerType: CFNumberType = 15;
+    pub(crate) const kCFNumberCFIndexType: CFNumberType = 14;
+    pub(crate) const kCFNumberDoubleType: CFNumberType = 13;
+    pub(crate) const kCFNumberFloatType: CFNumberType = 12;
+    pub(crate) const kCFNumberLongLongType: CFNumberType = 11;
+    pub(crate) const kCFNumberLongType: CFNumberType = 10;
+    pub(crate) const kCFNumberIntType: CFNumberType = 9;
+    pub(crate) const kCFNumberShortType: CFNumberType = 8;
+    pub(crate) const kCFNumberCharType: CFNumberType = 7;
+    pub(crate) const kCFNumberFloat64Type: CFNumberType = 6;
+    pub(crate) const kCFNumberFloat32Type: CFNumberType = 5;
+    pub(crate) const kCFNumberSInt64Type: CFNumberType = 4;
+    pub(crate) const kCFNumberSInt32Type: CFNumberType = 3;
+    pub(crate) const kCFNumberSInt16Type: CFNumberType = 2;
+    pub(crate) const kCFNumberSInt8Type: CFNumberType = 1;
+
+    pub(crate) const kCFStringEncodingUTF32LE: CFStringEncoding = 469762304;
+    pub(crate) const kCFStringEncodingUTF32BE: CFStringEncoding = 402653440;
+    pub(crate) const kCFStringEncodingUTF32: CFStringEncoding = 201326848;
+    pub(crate) const kCFStringEncodingUTF16LE: CFStringEncoding = 335544576;
+    pub(crate) const kCFStringEncodingUTF16BE: CFStringEncoding = 268435712;
+    pub(crate) const kCFStringEncodingUTF16: CFStringEncoding = 256;
+    pub(crate) const kCFStringEncodingNonLossyASCII: CFStringEncoding = 3071;
+    pub(crate) const kCFStringEncodingUnicode: CFStringEncoding = 256;
+    pub(crate) const kCFStringEncodingASCII: CFStringEncoding = 1536;
+    pub(crate) const kCFStringEncodingNextStepLatin: CFStringEncoding = 2817;
+    pub(crate) const kCFStringEncodingISOLatin1: CFStringEncoding = 513;
+    pub(crate) const kCFStringEncodingWindowsLatin1: CFStringEncoding = 1280;
+    pub(crate) const kCFStringEncodingMacRoman: CFStringEncoding = 0;
+
+    pub(crate) const kCFCompareForcedOrdering: CFStringCompareFlags = 512;
+    pub(crate) const kCFCompareWidthInsensitive: CFStringCompareFlags = 256;
+    pub(crate) const kCFCompareDiacriticInsensitive: CFStringCompareFlags = 128;
+    pub(crate) const kCFCompareNumerically: CFStringCompareFlags = 64;
+    pub(crate) const kCFCompareLocalized: CFStringCompareFlags = 32;
+    pub(crate) const kCFCompareNonliteral: CFStringCompareFlags = 16;
+    pub(crate) const kCFCompareAnchored: CFStringCompareFlags = 8;
+    pub(crate) const kCFCompareBackwards: CFStringCompareFlags = 4;
+    pub(crate) const kCFCompareCaseInsensitive: CFStringCompareFlags = 1;
+
+    // The CFArray wrapper is not mutable, so we use the APIs directly
+    pub(crate) type CFMutableArrayRef = *mut __CFArray;
+    extern "C" {
+        #[no_mangle]
+        pub(crate) fn CFArrayCreateMutable(
+            allocator: CFAllocatorRef,
+            capacity: CFIndex,
+            callBacks: *const CFArrayCallBacks,
+        ) -> CFMutableArrayRef;
+        #[no_mangle]
+        pub(crate) fn CFArrayAppendValue(theArray: CFMutableArrayRef, value: *const libc::c_void);
+    }
+    extern "C" {
+        // Missing
+        #[no_mangle]
+        pub(crate) fn CFStringCreateWithCStringNoCopy(
+            alloc: CFAllocatorRef,
+            cStr: *const libc::c_char,
+            encoding: CFStringEncoding,
+            contentsDeallocator: CFAllocatorRef,
+        ) -> CFStringRef;
+        #[no_mangle]
+        pub(crate) fn CFStringCompare(
+            theString1: CFStringRef,
+            theString2: CFStringRef,
+            compareOptions: CFStringCompareFlags,
+        ) -> CFComparisonResult;
+    }
+    extern "C" {
+        #[no_mangle]
+        pub(crate) fn CFDictionaryGetValue(
+            theDict: CFDictionaryRef,
+            key: *const libc::c_void,
+        ) -> *const libc::c_void;
+    }
+
+    // CFComparisonResult is missing PartialEq
+    pub(crate) fn comparison_was(a: CFComparisonResult, b: CFComparisonResult) -> bool {
+        match (a, b) {
+            (CFComparisonResult::LessThan, CFComparisonResult::LessThan) => true,
+            (CFComparisonResult::EqualTo, CFComparisonResult::EqualTo) => true,
+            (CFComparisonResult::GreaterThan, CFComparisonResult::GreaterThan) => true,
+            _ => false,
+        }
+    }
+
+    pub(crate) use core_graphics::{
+        base::CGFloat,
+        color::{CGColor, SysCGColorRef as CGColorRef},
+        font::CGGlyph,
+        geometry::{CGAffineTransform, CGPoint, CGRect, CGSize},
+    };
+    extern "C" {
+        #[no_mangle]
+        pub(crate) static CGAffineTransformIdentity: CGAffineTransform;
+        #[no_mangle]
+        pub(crate) fn CGRectIsNull(rect: CGRect) -> bool;
+    }
+    pub(crate) use core_text::run::CTRunRef;
+    extern "C" {
+        #[no_mangle]
+        pub(crate) fn CTRunGetGlyphCount(run: CTRunRef) -> CFIndex;
+        #[no_mangle]
+        pub(crate) fn CTRunGetAttributes(run: CTRunRef) -> CFDictionaryRef;
+        #[no_mangle]
+        pub(crate) fn CTRunGetGlyphs(run: CTRunRef, range: CFRange, buffer: *mut CGGlyph);
+        #[no_mangle]
+        pub(crate) fn CTRunGetPositions(run: CTRunRef, range: CFRange, buffer: *mut CGPoint);
+        #[no_mangle]
+        pub(crate) fn CTRunGetAdvances(run: CTRunRef, range: CFRange, buffer: *mut CGSize);
+        #[no_mangle]
+        pub(crate) fn CTLineGetGlyphCount(line: CTLineRef) -> CFIndex;
+        #[no_mangle]
+        pub(crate) fn CTLineGetGlyphRuns(line: CTLineRef) -> CFArrayRef;
+        #[no_mangle]
+        pub(crate) fn CTRunGetTypographicBounds(
+            run: CTRunRef,
+            range: CFRange,
+            ascent: *mut CGFloat,
+            descent: *mut CGFloat,
+            leading: *mut CGFloat,
+        ) -> libc::c_double;
+    }
+
+    pub(crate) use core_text::{
+        font::CTFontRef,
+        font_descriptor::{
+            kCTFontCascadeListAttribute, kCTFontDisplayNameAttribute, kCTFontFamilyNameAttribute,
+            kCTFontFeatureSettingsAttribute, kCTFontNameAttribute, kCTFontOrientationAttribute,
+            kCTFontURLAttribute, CTFontDescriptor, CTFontDescriptorCopyAttribute,
+            CTFontDescriptorCreateCopyWithAttributes,
+            CTFontDescriptorCreateMatchingFontDescriptors, CTFontDescriptorCreateWithAttributes,
+            CTFontDescriptorCreateWithNameAndSize, CTFontDescriptorRef, CTFontOrientation,
+        },
+        line::CTLineRef,
+        string_attributes::{
+            kCTFontAttributeName, kCTForegroundColorAttributeName, kCTKernAttributeName,
+            kCTVerticalFormsAttributeName,
+        },
+    };
+
+    pub(crate) const kCTFontVerticalOrientation: CTFontOrientation = 2;
+    pub(crate) const kCTFontHorizontalOrientation: CTFontOrientation = 1;
+    pub(crate) const kCTFontDefaultOrientation: CTFontOrientation = 0;
+    pub(crate) const kCTFontOrientationVertical: CTFontOrientation = 2;
+    pub(crate) const kCTFontOrientationHorizontal: CTFontOrientation = 1;
+    pub(crate) const kCTFontOrientationDefault: CTFontOrientation = 0;
+
+    // The CGFont wrapper is not feature complete.
+    pub(crate) type CGFontRef = *const __CGFont;
+    extern "C" {
+        pub(crate) type __CGFont;
+        #[no_mangle]
+        pub(crate) fn CGFontGetNumberOfGlyphs(font: CGFontRef) -> usize;
+        #[no_mangle]
+        pub(crate) fn CGFontRelease(font: CGFontRef);
+        #[no_mangle]
+        pub(crate) fn CGFontCopyGlyphNameForGlyph(font: CGFontRef, glyph: CGGlyph) -> CFStringRef;
+        #[no_mangle]
+        pub(crate) fn CTFontCopyGraphicsFont(
+            font: CTFontRef,
+            attributes: *mut CTFontDescriptorRef,
+        ) -> CGFontRef;
+        #[no_mangle]
+        pub(crate) static kCTFontPostScriptNameKey: CFStringRef;
+    }
+    // Typesetters
+    pub(crate) type CTTypesetterRef = *const __CTTypesetter;
+    extern "C" {
+        pub(crate) type __CTTypesetter;
+        #[no_mangle]
+        pub(crate) fn CTTypesetterCreateWithAttributedString(
+            string: CFAttributedStringRef,
+        ) -> CTTypesetterRef;
+        #[no_mangle]
+        pub(crate) fn CTTypesetterCreateLine(
+            typesetter: CTTypesetterRef,
+            stringRange: CFRange,
+        ) -> CTLineRef;
+    }
+
+    // misc
+    extern "C" {
+        #[no_mangle]
+        pub(crate) fn CTFontCreateWithFontDescriptor(
+            descriptor: CTFontDescriptorRef,
+            size: CGFloat,
+            matrix: *const CGAffineTransform,
+        ) -> CTFontRef;
+        #[no_mangle]
+        pub(crate) fn CTLineCreateJustifiedLine(
+            line: CTLineRef,
+            justificationFactor: CGFloat,
+            justificationWidth: libc::c_double,
+        ) -> CTLineRef;
+        #[no_mangle]
+        pub(crate) fn CFStringCreateWithCharactersNoCopy(
+            alloc: CFAllocatorRef,
+            chars: *const UniChar,
+            numChars: CFIndex,
+            contentsDeallocator: CFAllocatorRef,
+        ) -> CFStringRef;
+        #[no_mangle]
+        pub(crate) fn CTFontCreateCopyWithAttributes(
+            font: CTFontRef,
+            size: CGFloat,
+            matrix: *const CGAffineTransform,
+            attributes: CTFontDescriptorRef,
+        ) -> CTFontRef;
+        #[no_mangle]
+        pub(crate) fn CTFontCopyAttribute(font: CTFontRef, attribute: CFStringRef) -> CFTypeRef;
+        #[no_mangle]
+        pub(crate) fn CTFontCopyName(font: CTFontRef, nameKey: CFStringRef) -> CFStringRef;
+        #[no_mangle]
+        pub(crate) fn CTFontCopyLocalizedName(
+            font: CTFontRef,
+            nameKey: CFStringRef,
+            actualLanguage: *mut CFStringRef,
+        ) -> CFStringRef;
+        #[no_mangle]
+        pub(crate) fn CTFontGetGlyphsForCharacters(
+            font: CTFontRef,
+            characters: *const UniChar,
+            glyphs: *mut CGGlyph,
+            count: CFIndex,
+        ) -> bool;
+        #[no_mangle]
+        pub(crate) fn CTFontGetGlyphWithName(font: CTFontRef, glyphName: CFStringRef) -> CGGlyph;
+        #[no_mangle]
+        pub(crate) fn CTFontGetBoundingRectsForGlyphs(
+            font: CTFontRef,
+            orientation: CTFontOrientation,
+            glyphs: *const CGGlyph,
+            boundingRects: *mut CGRect,
+            count: CFIndex,
+        ) -> CGRect;
+        #[no_mangle]
+        pub(crate) fn CTFontGetAdvancesForGlyphs(
+            font: CTFontRef,
+            orientation: CTFontOrientation,
+            glyphs: *const CGGlyph,
+            advances: *mut CGSize,
+            count: CFIndex,
+        ) -> libc::c_double;
+        #[no_mangle]
+        pub(crate) static kCTFontFeatureTypeIdentifierKey: CFStringRef;
+        #[no_mangle]
+        pub(crate) static kCTFontFeatureTypeNameKey: CFStringRef;
+        #[no_mangle]
+        pub(crate) static kCTFontFeatureTypeSelectorsKey: CFStringRef;
+        #[no_mangle]
+        pub(crate) static kCTFontFeatureTypeExclusiveKey: CFStringRef;
+        #[no_mangle]
+        pub(crate) fn CTFontCopyFeatures(font: CTFontRef) -> CFArrayRef;
+        #[no_mangle]
+        pub(crate) static kCTFontFeatureSelectorNameKey: CFStringRef;
+        #[no_mangle]
+        pub(crate) static kCTFontFeatureSelectorDefaultKey: CFStringRef;
+        #[no_mangle]
+        pub(crate) static kCTFontFeatureSelectorIdentifierKey: CFStringRef;
+
+        #[no_mangle]
+        pub(crate) static kCTFontFullNameKey: CFStringRef;
+        #[no_mangle]
+        pub(crate) static kCTFontStyleNameKey: CFStringRef;
+        #[no_mangle]
+        pub(crate) static kCTFontFamilyNameKey: CFStringRef;
+
+        #[no_mangle]
+        pub(crate) fn CTFontGetXHeight(font: CTFontRef) -> CGFloat;
+        #[no_mangle]
+        pub(crate) fn CTFontGetAscent(font: CTFontRef) -> CGFloat;
+        #[no_mangle]
+        pub(crate) fn CTFontGetDescent(font: CTFontRef) -> CGFloat;
+        #[no_mangle]
+        pub(crate) fn CTFontGetGlyphCount(font: CTFontRef) -> CFIndex;
+        #[no_mangle]
+        pub(crate) fn CTFontGetSlantAngle(font: CTFontRef) -> CGFloat;
+        #[no_mangle]
+        pub(crate) fn CTFontGetCapHeight(font: CTFontRef) -> CGFloat;
+        #[no_mangle]
+        pub(crate) fn CFStringGetCharacters(
+            theString: CFStringRef,
+            range: CFRange,
+            buffer: *mut UniChar,
+        );
+        #[no_mangle]
+        pub(crate) fn CFArrayGetValueAtIndex(
+            theArray: CFArrayRef,
+            idx: CFIndex,
+        ) -> *const libc::c_void;
+        #[no_mangle]
+        pub(crate) fn CTFontGetSize(font: CTFontRef) -> CGFloat;
+        #[no_mangle]
+        pub(crate) fn CTFontGetMatrix(font: CTFontRef) -> CGAffineTransform;
+        #[no_mangle]
+        pub(crate) fn CGColorGetComponents(color: CGColorRef) -> *const CGFloat;
+        #[no_mangle]
+        pub(crate) fn CFBooleanGetValue(boolean: CFBooleanRef) -> libc::c_uchar;
+    }
+
+    pub(crate) type UniChar = UInt16;
+    pub(crate) type UInt16 = libc::c_ushort;
+
+    #[inline(always)]
+    pub(crate) fn CFRangeMake(mut loc: CFIndex, mut len: CFIndex) -> CFRange {
+        let mut range: CFRange = CFRange {
+            location: 0,
+            length: 0,
+        };
+        range.location = loc;
+        range.length = len;
+        range
+    }
+
+    pub(crate) unsafe fn cgColorToRGBA32(mut color: CGColorRef) -> u32 {
+        let mut components: *const CGFloat = CGColorGetComponents(color);
+        let mut rval: u32 = (*components.offset(0) * 255.0f64 + 0.5f64) as u8 as u32;
+        rval <<= 8i32;
+        rval = (rval as u32).wrapping_add((*components.offset(1) * 255.0f64 + 0.5f64) as u8 as u32);
+        rval <<= 8i32;
+        rval = (rval as u32).wrapping_add((*components.offset(2) * 255.0f64 + 0.5f64) as u8 as u32);
+        rval <<= 8i32;
+        rval = (rval as u32).wrapping_add((*components.offset(3) * 255.0f64 + 0.5f64) as u8 as u32);
+        return rval;
+    }
+}
